Markups following feedback from AJH
